re-moving files to UNUSED folder
