+++ src/components/pages/UNUSED/ProfileList/ProfileListContainer.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { useOktaAuth } from '@okta/okta-react';
 
-import { getProfileData } from '../../../api';
+import { getProfileData } from '../../../../api';
 
-import { List } from '../../common';
+import { List } from '../../../common';
 
 import RenderProfileListPage from './RenderProfileListPage';
